#58,#63 added game object generation  Game object gets inserted to database when first player joins the lobby
