+++ frontend/views/game/index.html
@@ -13,6 +13,6 @@
       <canvas id="gameCanvas"></canvas>
     </main>
 
-    <script type="module" src="/index.js"></script>
+    <script type="module" src="./index.js"></script>
   </body>
 </html>
